:truck: Move chain logic to blocks
Remove unwanted chain.ts
